crates/graph: separate API for manipulating graph from code building it
